% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{ap_at_n}
\alias{ap_at_n}
\title{Average Precision at n.
n is determined by the boundaries parameter.
boundaries parameter means the cutoff points according to the total results.}
\usage{
ap_at_n(interactions, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of AP at n calculation based on the boundaries parameter.
}
\description{
Average Precision at n.
n is determined by the boundaries parameter.
boundaries parameter means the cutoff points according to the total results.
}
