% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{bpref}
\alias{bpref}
\title{BPref.
boundaries parameter means the cutoff points according to the total results.}
\usage{
bpref(interactions, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of BPref calculation based on the boundaries parameter.
}
\description{
BPref.
boundaries parameter means the cutoff points according to the total results.
}
