% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{cgain}
\alias{cgain}
\title{Cumulative Gain.
boundaries parameter means the cutoff points according to the total results.}
\usage{
cgain(interactions, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of CG calculation based on the boundaries parameter.
}
\description{
Cumulative Gain.
boundaries parameter means the cutoff points according to the total results.
}
