% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{dcgain}
\alias{dcgain}
\title{Discounted Cumulative Gain.
boundaries parameter means the cutoff points according to the total results.}
\usage{
dcgain(interactions, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of DCG calculation based on the boundaries parameter.
}
\description{
Discounted Cumulative Gain.
boundaries parameter means the cutoff points according to the total results.
}
