% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{err}
\alias{err}
\title{Expected Reciprocal Rank.
boundaries parameter means the cutoff points according to the total results.
max_grade parameter is used in the calculation of ERR.
The minimum level is also used in the calculation but does not need to be included as a parameter for this measurement.}
\usage{
err(interactions, max_grade = 5, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{max_grade}{An integer.}

\item{boundaries}{A list.}
}
\value{
The results of ERR calculation based on the boundaries parameter.
}
\description{
Expected Reciprocal Rank.
boundaries parameter means the cutoff points according to the total results.
max_grade parameter is used in the calculation of ERR.
The minimum level is also used in the calculation but does not need to be included as a parameter for this measurement.
}
