% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{fmeasure}
\alias{fmeasure}
\title{F-Measure.
boundaries parameter means the cutoff points according to the total results.}
\usage{
fmeasure(interactions, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of F-Measure calculation based on the boundaries parameter.
}
\description{
F-Measure.
boundaries parameter means the cutoff points according to the total results.
}
