% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{gmap}
\alias{gmap}
\title{Geometric Mean Average Precision.
boundaries parameter means the cutoff points according to the total results.
constant parameter is used in the calculation of GMAP in order to avoid zero values during the calculation.}
\usage{
gmap(interactions, constant = 0.01, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{constant}{A double.}

\item{boundaries}{A list.}
}
\value{
The results of GMAP calculation based on the boundaries parameter.
}
\description{
Geometric Mean Average Precision.
boundaries parameter means the cutoff points according to the total results.
constant parameter is used in the calculation of GMAP in order to avoid zero values during the calculation.
}
