% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{iap}
\alias{iap}
\title{Eleven Point - Interpolated Average Precision.}
\usage{
iap(interactions)
}
\arguments{
\item{interactions}{A list.}
}
\value{
The results of IAP calculation.
}
\description{
Eleven Point - Interpolated Average Precision.
}
