% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{mean_ap}
\alias{mean_ap}
\title{Mean Average Precision.
boundaries parameter means the cutoff points according to the total results.}
\usage{
mean_ap(interactions, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of MAP calculation based on the boundaries parameter.
}
\description{
Mean Average Precision.
boundaries parameter means the cutoff points according to the total results.
}
