% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{mrr}
\alias{mrr}
\title{Mean Reciprocal Rank.}
\usage{
mrr(interactions)
}
\arguments{
\item{interactions}{A list.}
}
\value{
The results of MRR calculation.
}
\description{
Mean Reciprocal Rank.
}
