% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{ncgain}
\alias{ncgain}
\title{Normalized Cumulative Gain.
boundaries parameter means the cutoff points according to the total results.}
\usage{
ncgain(interactions, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of NCG calculation based on the boundaries parameter.
}
\description{
Normalized Cumulative Gain.
boundaries parameter means the cutoff points according to the total results.
}
