% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{ndcgain}
\alias{ndcgain}
\title{Normalized Discounted Cumulative Gain.
boundaries parameter means the cutoff points according to the total results.}
\usage{
ndcgain(interactions, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of NDCG calculation based on the boundaries parameter.
}
\description{
Normalized Discounted Cumulative Gain.
boundaries parameter means the cutoff points according to the total results.
}
