% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{rbprecision}
\alias{rbprecision}
\title{Rank-Biased Precision.
boundaries parameter means the cutoff points according to the total results.
p --persistence (or probability) levels-- parameter is used in the calculation of RBP.}
\usage{
rbprecision(interactions, p = list(0.5, 0.8, 0.95),
  boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{p}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of RBP calculation based on the boundaries parameter.
}
\description{
Rank-Biased Precision.
boundaries parameter means the cutoff points according to the total results.
p --persistence (or probability) levels-- parameter is used in the calculation of RBP.
}
