% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{rprecision}
\alias{rprecision}
\title{R-Precision.
boundaries parameter means the cutoff points according to the total results.}
\usage{
rprecision(interactions, boundaries = list("all"))
}
\arguments{
\item{interactions}{A list.}

\item{boundaries}{A list.}
}
\value{
The results of R-Precision calculation based on the boundaries parameter.
}
\description{
R-Precision.
boundaries parameter means the cutoff points according to the total results.
}
